ZLib work around.
It seems ZLib wants a two byte header (constant at 0x78, 0xDA) and a four
byte trailer (the Adler32 of the uncompressed data).  The two byte header is
easy to reconstruct, since it's constant, but the Adler32 isn't quite that
easy to get before we uncompress the data.  So, we add it to the Lua ZIP
extention data as a way to sneak it through (a gross hack).

I'm still trying to get it working, and in theory, it should, but while in
theory, there is no difference between theory and practice, in practice ...