Include a language interpreter in the extra file information.
The thought behind this is---there are Lua modules that are specific to
LuaJIT, and some that are specific to environments that include Lua but
aren't specifically Lua (if you get my drift, like World of Warcraft).  So
I thought I would include a field for the interpreter.  But if I'm doing that,
I might as well make it such that any scripting langauge can use this.